+++ library/src/main/java/com/codetroopers/betterpickers/hmspicker/HmsPickerBuilder.java
@@ -5,8 +5,6 @@ import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
 import android.util.Log;
 
-import com.codetroopers.betterpickers.hmspicker.HmsPickerDialogFragment.HmsPickerDialogHandler;
-
 import java.util.Vector;
 
 /**
@@ -18,10 +16,26 @@ public class HmsPickerBuilder {
     private Integer styleResId; // Required
     private Fragment targetFragment;
     private int mReference;
-    private Vector<HmsPickerDialogHandler> mHmsPickerDialogHandlers = new Vector<HmsPickerDialogHandler>();
+    private Vector<HmsPickerDialogFragment.HmsPickerDialogHandlerV2> mHmsPickerDialogHandlerV2s = new Vector<HmsPickerDialogFragment.HmsPickerDialogHandlerV2>();
     private int mHours;
     private int mMinutes;
     private int mSeconds;
+    private Integer plusMinusVisibility;
+
+    /**
+     * Set the visibility of the +/- button. This takes an int corresponding to Android's View.VISIBLE, View.INVISIBLE,
+     * or View.GONE.  When using View.INVISIBLE, the +/- button will still be present in the layout but be
+     * non-clickable. When set to View.GONE, the +/- button will disappear entirely, and the "0" button will occupy its
+     * space.
+     *
+     * @param plusMinusVisibility an int corresponding to View.VISIBLE, View.INVISIBLE, or View.GONE
+     * @return the current Builder object
+     */
+    public HmsPickerBuilder setPlusMinusVisibility(int plusMinusVisibility) {
+        this.plusMinusVisibility = plusMinusVisibility;
+        return this;
+    }
+
 
     /**
      * Attach a FragmentManager. This is required for creation of the Fragment.
@@ -76,8 +90,8 @@ public class HmsPickerBuilder {
      * @param handler an Object implementing the appropriate Picker Handler
      * @return the current Builder object
      */
-    public HmsPickerBuilder addHmsPickerDialogHandler(HmsPickerDialogHandler handler) {
-        this.mHmsPickerDialogHandlers.add(handler);
+    public HmsPickerBuilder addHmsPickerDialogHandler(HmsPickerDialogFragment.HmsPickerDialogHandlerV2 handler) {
+        this.mHmsPickerDialogHandlerV2s.add(handler);
         return this;
     }
 
@@ -87,15 +101,15 @@ public class HmsPickerBuilder {
      * @param handler the Object to remove
      * @return the current Builder object
      */
-    public HmsPickerBuilder removeHmsPickerDialogHandler(HmsPickerDialogHandler handler) {
-        this.mHmsPickerDialogHandlers.remove(handler);
+    public HmsPickerBuilder removeHmsPickerDialogHandler(HmsPickerDialogFragment.HmsPickerDialogHandlerV2 handler) {
+        this.mHmsPickerDialogHandlerV2s.remove(handler);
         return this;
     }
 
     /**
      * Set some initial values for the picker
      *
-     * @param hours the initial hours value
+     * @param hours   the initial hours value
      * @param minutes the initial minutes value
      * @param seconds the initial seconds value
      * @return the current Builder object
@@ -140,18 +154,19 @@ public class HmsPickerBuilder {
             Log.e("HmsPickerBuilder", "setFragmentManager() and setStyleResId() must be called.");
             return;
         }
-        final FragmentTransaction ft = manager.beginTransaction();
+        FragmentTransaction ft = manager.beginTransaction();
         final Fragment prev = manager.findFragmentByTag("hms_dialog");
         if (prev != null) {
-            ft.remove(prev);
+            ft.remove(prev).commit();
+            ft = manager.beginTransaction();
         }
         ft.addToBackStack(null);
 
-        final HmsPickerDialogFragment fragment = HmsPickerDialogFragment.newInstance(mReference, styleResId);
+        final HmsPickerDialogFragment fragment = HmsPickerDialogFragment.newInstance(mReference, styleResId, plusMinusVisibility);
         if (targetFragment != null) {
             fragment.setTargetFragment(targetFragment, 0);
         }
-        fragment.setHmsPickerDialogHandlers(mHmsPickerDialogHandlers);
+        fragment.setHmsPickerDialogHandlersV2(mHmsPickerDialogHandlerV2s);
 
         if ((mHours | mMinutes | mSeconds) != 0) {
             fragment.setTime(mHours, mMinutes, mSeconds);